+++ src/Mcfedr/JsonFormBundle/Exception/JsonHttpException.php
@@ -1,17 +1,22 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Mcfedr\JsonFormBundle\Exception;
 
 use Symfony\Component\HttpKernel\Exception\HttpException;
 
 abstract class JsonHttpException extends HttpException
 {
+    /**
+     * @var null|mixed
+     */
     private $data;
 
     /**
-     * @param int    $statusCode
-     * @param string $message
-     * @param mixed  $data
+     * @param int        $statusCode
+     * @param string     $message
+     * @param null|mixed $data
      */
     public function __construct($statusCode, $message = null, $data = null)
     {
@@ -20,18 +25,12 @@ abstract class JsonHttpException extends HttpException
         $this->setData($data);
     }
 
-    /**
-     * @return mixed
-     */
     public function getData()
     {
         return $this->data;
     }
 
-    /**
-     * @param mixed $data
-     */
-    public function setData($data)
+    public function setData($data): void
     {
         $this->data = $data;
     }