+++ src/Mcfedr/JsonFormBundle/Exception/MissingFormHttpException.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Mcfedr\JsonFormBundle\Exception;
 
 use Symfony\Component\Form\FormInterface;
@@ -8,6 +10,6 @@ class MissingFormHttpException extends JsonHttpException
 {
     public function __construct(FormInterface $form)
     {
-        parent::__construct(400, 'Missing ' . $form->getName());
+        parent::__construct(400, 'Missing '.$form->getName());
     }
 }