+++ tests/Mcfedr/JsonFormBundle/Tests/Controller/JsonControllerCheckboxTest.php
@@ -1,8 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Mcfedr\JsonFormBundle\Tests\Controller;
 
 use Mcfedr\JsonFormBundle\Controller\JsonController;
+use PHPUnit\Framework\MockObject\MockObject;
+use PHPUnit\Framework\TestCase;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Validator\ValidatorExtension;
 use Symfony\Component\Form\Form;
@@ -10,10 +14,13 @@ use Symfony\Component\Form\Forms;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Validator\Validation;
 
-class JsonControllerCheckboxTest extends \PHPUnit_Framework_TestCase
+/**
+ * @internal
+ */
+final class JsonControllerCheckboxTest extends TestCase
 {
     /**
-     * @var JsonController
+     * @var MockObject^JsonController
      */
     private $controller;
 
@@ -22,42 +29,43 @@ class JsonControllerCheckboxTest extends \PHPUnit_Framework_TestCase
      */
     private $form;
 
-    public function setUp()
+    protected function setUp(): void
     {
-        $this->controller = $this->getMockForAbstractClass('Mcfedr\JsonFormBundle\Controller\JsonController');
+        $this->controller = $this->getMockBuilder(JsonController::class)->getMock();
         $validator = Validation::createValidator();
         $this->form = Forms::createFormFactoryBuilder()
             ->addExtension(new ValidatorExtension($validator))
             ->getFormFactory()
             ->createBuilder()
             ->add('one', CheckboxType::class)
-            ->getForm();
+            ->getForm()
+        ;
     }
 
-    public function testHandleTrue()
+    public function testHandleTrue(): void
     {
         $request = new Request([], [], [], [], [], [], json_encode([
             'form' => [
-                'one' => true
-            ]
+                'one' => true,
+            ],
         ]));
 
         $this->invokeMethod($this->controller, 'handleJsonForm', [$this->form, $request]);
 
-        $this->assertEquals($this->form->getData()['one'], true);
+        static::assertEquals($this->form->getData()['one'], true);
     }
 
-    public function testHandleFalse()
+    public function testHandleFalse(): void
     {
         $request = new Request([], [], [], [], [], [], json_encode([
             'form' => [
-                'one' => false
-            ]
+                'one' => false,
+            ],
         ]));
 
         $this->invokeMethod($this->controller, 'handleJsonForm', [$this->form, $request]);
 
-        $this->assertEquals($this->form->getData()['one'], false);
+        static::assertEquals($this->form->getData()['one'], false);
     }
 
     /**
@@ -71,7 +79,7 @@ class JsonControllerCheckboxTest extends \PHPUnit_Framework_TestCase
      */
     private function invokeMethod($object, $methodName, array $parameters = [])
     {
-        $reflection = new \ReflectionClass(get_class($object));
+        $reflection = new \ReflectionClass(\get_class($object));
         $method = $reflection->getMethod($methodName);
         $method->setAccessible(true);