+++ tests/Mcfedr/JsonFormBundle/Type/TestType.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Mcfedr\JsonFormBundle\Type;
 
 use Symfony\Component\Form\AbstractType;
@@ -12,17 +14,18 @@ use Symfony\Component\Validator\Constraints\Email;
 
 class TestType extends AbstractType
 {
-    public function buildForm(FormBuilderInterface $builder, array $options)
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
             ->add('one', ChoiceType::class, [
-                'choices' => ['value' => 'value']
+                'choices' => ['value' => 'value'],
             ])
             ->add('two', CheckboxType::class)
             ->add('number', NumberType::class)
             ->add('email', EmailType::class, [
                 'constraints' => new Email(),
-            ]);
+            ])
+        ;
     }
 
     /**
@@ -30,7 +33,7 @@ class TestType extends AbstractType
      *
      * @return string The name of this type
      */
-    public function getBlockPrefix()
+    public function getBlockPrefix(): string
     {
         return 'form';
     }