+++ tests/bootstrap.php
@@ -1,6 +1,8 @@
 <?php
 
-$loader = require __DIR__ . '/../vendor/autoload.php';
+declare(strict_types=1);
+
+$loader = require __DIR__.'/../vendor/autoload.php';
 \Doctrine\Common\Annotations\AnnotationRegistry::registerLoader([$loader, 'loadClass']);
 
 return $loader;