+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 4.6.0
+- Drop support for php less than 7.2
+- Add strict_types
+- Upgrade phpunit
+- Add phpstan
+
+## 4.5.0
+- Fix use of getParameter in JsonControllerTrait
+- Deprecate JsonController
+
 ## 4.4.0
 - Change to using FormInterface instead of Form